Added tests for partition::PartitionSet.
Added addtional tests to directly cover all methods of PartitionSet and
consolidated all PartitionSet tests in the partition::set::test module.
